pyproject.toml
[tool.poetry]
name = "instead-of-me-store-backend"
version = "0.1.0"
description = ""
authors = ["lovers2020 <dkrkskswl@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
